Updates for arbitrary deck creation
Realized that it would be useful to be able to create one-off copies when we have to reschedule meetings because of one off occurrences such as stat holidays.

Created methods to do that in testSlideCreation, and refactored the code used to generate the slides so that we have a generic slide creation function that doesn't do any date or sprint string calculations.